cyw43455: add FWIL layer — cyw43455_fwil.c
Introduce cyw43455_fwil.c as the bus-agnostic IOVAR/IOCTL encoding
layer between callers (cfg, scan, security) and the SDPCM TX/RX
primitives.

- Move cyw_sdpcm_recv_one (static) and cyw_iovar_get from cyw43455_fw.c
  into fwil.c; generalise into cyw_fil_txrx (internal) + public API:
    cyw_fil_iovar_data_get / _set
    cyw_fil_iovar_int_get / _set
    cyw_fil_cmd_data_get / _set
- cyw43455_fw.c: replace cyw_iovar_get("ver") call with
  cyw_fil_iovar_data_get; remove now-dead ALIGN4 macro.
- cyw43455_var.h: update declarations; add WLC_UP/DOWN and Milestone 2
  raw-command codes needed by cfg.c.
- Makefile.cyw43455: add cyw43455_fwil.c to SRCS.
