+++ cyw43455_fw.c
@@ -35,9 +35,6 @@
 
 #define CYW_CLM_CHUNK	16384	/* bytes per clmload iovar call */
 
-/* Align a size up to the nearest 4 bytes */
-#define ALIGN4(x) (((x) + 3) & ~3)
-
 /* -------------------------------------------------------------------------
  * NVRAM parsing
  *
@@ -117,240 +114,6 @@ cyw_parse_nvram(const uint8_t *raw, size_t rawlen, size_t *outlen)
 
 	*outlen = buflen;
 	return (buf);
-}
-
-/* -------------------------------------------------------------------------
- * BCDC IOVAR helpers
- * ------------------------------------------------------------------------- */
-
-/*
- * cyw_sdpcm_recv_one — read one SDPCM frame from the F2 FIFO.
- *
- * Gates on INTSTATUS I_HMB_SW_MASK; returns EAGAIN if no frame is indicated.
- * On a positive gate, clears the bits and reads the full frame (rounded up
- * to CYW_F2_BLKSIZE).  Updates sc->sdpcm_rx_max from hdr->credit.
- */
-static int
-cyw_sdpcm_recv_one(struct cyw_softc *sc, uint8_t *buf, uint16_t *out_flen)
-{
-	device_t parent = device_get_parent(sc->dev);
-	struct cyw_sdpcm_hdr *hdr;
-	uint16_t flen;
-	size_t rdlen;
-	int err;
-
-	if (sc->sdio_core_base != 0) {
-		uint32_t intstatus = cyw_bp_read32(sc,
-		    sc->sdio_core_base + SD_REG_INTSTATUS);
-		if ((intstatus & (I_HMB_SW_MASK | I_XMTDATA_AVAIL)) == 0)
-			return (EAGAIN);
-		/* Clear only the soft-mailbox bits; I_XMTDATA_AVAIL auto-clears. */
-		if (intstatus & I_HMB_SW_MASK)
-			cyw_bp_write32(sc, sc->sdio_core_base + SD_REG_INTSTATUS,
-			    intstatus & I_HMB_SW_MASK);
-	}
-
-	/*
-	 * Peek: read enough to get the SDPCM frame length (first 2 bytes),
-	 * using 64 bytes so the SDHCI block size matches F2's block size.
-	 */
-	err = SDIO_READ_EXTENDED(parent, 2 /* F2 */, CYW_F2_FIFO_ADDR,
-	    CYW_F2_BLKSIZE, buf, false /* FIFO, fixed addr */);
-	if (err)
-		return (err);
-
-	hdr = (struct cyw_sdpcm_hdr *)buf;
-	flen = le16toh(hdr->len);
-
-	/* Linux brcmf_sdio_hdparse: all-zero means no more frames. */
-	if (flen == 0 && le16toh(hdr->len_inv) == 0)
-		return (EAGAIN);
-
-	/* HW checksum: (uint16_t)~(flen ^ ~flen) must be 0. */
-	if ((uint16_t)~(flen ^ le16toh(hdr->len_inv)) != 0)
-		return (EAGAIN);
-
-	if (flen == 0xFFFF || flen < CYW_SDPCM_HDR_LEN)
-		return (EAGAIN);
-	if (flen > CYW_SDPCM_MAX_FRAME)
-		return (EINVAL);
-
-	if (flen <= CYW_F2_BLKSIZE) {
-		/* Frame fits in the block we already read. */
-		goto done;
-	}
-
-	/*
-	 * F2 is a streaming FIFO — the second CMD53 continues where the
-	 * first left off.  Read only the remainder (flen - 64) rounded up
-	 * to a block boundary, appending at buf+64 (not overwriting buf[0]).
-	 * Linux: brcmf_sdio_read_control() / brcmf_sdio_readframes() with
-	 * BRCMF_FIRSTREAD=64 do exactly this.
-	 */
-	rdlen = ((size_t)(flen - CYW_F2_BLKSIZE) + CYW_F2_BLKSIZE - 1) &
-	    ~(size_t)(CYW_F2_BLKSIZE - 1);
-	err = SDIO_READ_EXTENDED(parent, 2 /* F2 */, CYW_F2_FIFO_ADDR,
-	    rdlen, buf + CYW_F2_BLKSIZE, false /* FIFO, append after first block */);
-	if (err)
-		return (err);
-
-done:
-	CYW_LOCK(sc);
-	sc->sdpcm_rx_max = hdr->credit;
-	CYW_UNLOCK(sc);
-
-	if (out_flen != NULL)
-		*out_flen = flen;
-	return (0);
-}
-
-/*
- * cyw_iovar_get — send a BCDC GET IOVAR over F2, return response payload.
- *
- * Frame layout:
- *   [SDPCM header 12B][BCDC header 16B][name NUL-terminated][zero-padded buf]
- *
- * TX credit management: firmware grants credits via hdr->credit.  We must not
- * advance sdpcm_tx_seq past sdpcm_rx_max.  On first call sdpcm_rx_max==0 so
- * we drain RX frames to collect firmware's boot credit grant before sending.
- */
-int
-cyw_iovar_get(struct cyw_softc *sc, const char *name, void *buf, size_t buflen)
-{
-	struct cyw_sdpcm_hdr *sph;
-	struct cyw_bcdc_hdr  *bch;
-	size_t namelen = strlen(name) + 1;
-	size_t payload  = namelen + buflen;
-	size_t framelen = ALIGN4(CYW_SDPCM_HDR_LEN + CYW_BCDC_HDR_LEN + payload);
-	uint8_t *frame, *rsp;
-	uint16_t id;
-	int err, i;
-
-	frame = malloc(framelen, M_CYW43455, M_WAITOK | M_ZERO);
-	/* Extra CYW_F2_BLKSIZE: second read appends at buf+64, needs room. */
-	rsp   = malloc(CYW_SDPCM_MAX_FRAME + CYW_F2_BLKSIZE, M_CYW43455, M_WAITOK | M_ZERO);
-
-	id = ++sc->ioctl_id;
-
-	sph = (struct cyw_sdpcm_hdr *)frame;
-	sph->len         = htole16((uint16_t)framelen);
-	sph->len_inv     = htole16(~(uint16_t)framelen);
-	sph->seq         = sc->sdpcm_tx_seq;
-	sph->chan_flags  = CYW_SDPCM_CHAN_CTRL;
-	sph->data_offset = CYW_SDPCM_HDR_LEN;
-
-	bch = (struct cyw_bcdc_hdr *)(frame + CYW_SDPCM_HDR_LEN);
-	bch->cmd   = htole32(WLC_GET_VAR);
-	bch->len   = htole32((uint32_t)payload);
-	bch->flags = htole32(((uint32_t)id << BCDC_DCMD_ID_SHIFT));
-
-	memcpy(frame + CYW_SDPCM_HDR_LEN + CYW_BCDC_HDR_LEN, name, namelen);
-
-	/*
-	 * Drain RX until firmware grants at least one TX credit.
-	 * Firmware sends a boot SDPCM frame with credit > 0; without draining
-	 * it we would spin forever since sdpcm_rx_max starts at 0.
-	 */
-	for (i = 0; i < 100; i++) {
-		uint8_t delta = (uint8_t)(sc->sdpcm_rx_max - sc->sdpcm_tx_seq);
-		if (delta > 0)
-			break;
-		err = cyw_sdpcm_recv_one(sc, rsp, NULL);
-		if (err != 0 && err != EAGAIN) {
-			free(frame, M_CYW43455);
-			free(rsp, M_CYW43455);
-			return (err);
-		}
-		DELAY(10000);
-	}
-
-	if ((uint8_t)(sc->sdpcm_rx_max - sc->sdpcm_tx_seq) == 0) {
-		device_printf(sc->dev,
-		    "cyw_iovar_get '%s': no TX credits after drain\n", name);
-		free(frame, M_CYW43455);
-		free(rsp, M_CYW43455);
-		return (ENOBUFS);
-	}
-
-	sc->sdpcm_tx_seq++;
-
-	err = cyw_f2_write_block(sc, frame, framelen);
-	if (err) {
-		free(frame, M_CYW43455);
-		free(rsp, M_CYW43455);
-		return (err);
-	}
-
-	/*
-	 * Ring the doorbell: write SMB_HOST_INT to TOSBMAILBOX so firmware
-	 * wakes from SR and checks F2.  Without this, SR-capable chips never
-	 * process the frame we just wrote.
-	 */
-	if (sc->sdio_core_base != 0)
-		cyw_bp_write32(sc, sc->sdio_core_base + SD_REG_TOSBMAILBOX,
-		    SMB_HOST_INT);
-
-	if (sc->sdio_core_base != 0) {
-		uint32_t s = cyw_bp_read32(sc,
-		    sc->sdio_core_base + SD_REG_INTSTATUS);
-		device_printf(sc->dev,
-		    "cyw_iovar_get '%s': post-TX INTSTATUS=0x%08x\n", name, s);
-	}
-
-	/*
-	 * Poll F2 FIFO for the IOVAR response.  Blind 64-byte reads; EAGAIN
-	 * means FIFO empty.  Match on CTRL channel and BCDC transaction id.
-	 */
-	err = ETIMEDOUT;
-	for (i = 0; i < 300; i++) {
-		struct cyw_sdpcm_hdr *rsph;
-		struct cyw_bcdc_hdr  *rbch;
-		uint16_t flen, rsp_id;
-		int rerr;
-
-		DELAY(10000);
-
-		rerr = cyw_sdpcm_recv_one(sc, rsp, &flen);
-		if (rerr == EAGAIN)
-			continue;
-		if (rerr != 0) {
-			err = rerr;
-			break;
-		}
-
-		rsph = (struct cyw_sdpcm_hdr *)rsp;
-		if (flen < CYW_SDPCM_HDR_LEN + CYW_BCDC_HDR_LEN)
-			continue;
-		if ((rsph->chan_flags & 0x0f) != CYW_SDPCM_CHAN_CTRL)
-			continue;
-
-		rbch = (struct cyw_bcdc_hdr *)(rsp + rsph->data_offset);
-		rsp_id = (uint16_t)
-		    ((le32toh(rbch->flags) >> BCDC_DCMD_ID_SHIFT) & 0xffff);
-		if (rsp_id != id)
-			continue;
-		if (le32toh(rbch->flags) & BCDC_DCMD_ERROR) {
-			err = EIO;
-			break;
-		}
-
-		uint8_t *payload_ptr = (uint8_t *)rbch + CYW_BCDC_HDR_LEN;
-		size_t avail = (size_t)(rsp + flen - payload_ptr);
-		memcpy(buf, payload_ptr, MIN(buflen, avail));
-		err = 0;
-		break;
-	}
-
-	if (err == ETIMEDOUT) {
-		uint32_t s = (sc->sdio_core_base != 0) ?
-		    cyw_bp_read32(sc, sc->sdio_core_base + SD_REG_INTSTATUS) : 0;
-		device_printf(sc->dev,
-		    "cyw_iovar_get '%s': timed out INTSTATUS=0x%08x\n", name, s);
-	}
-
-	free(frame, M_CYW43455);
-	free(rsp, M_CYW43455);
-	return (err);
 }
 
 /* -------------------------------------------------------------------------
@@ -599,7 +362,7 @@ cyw_fw_download(struct cyw_softc *sc)
 
 	/* --- Read firmware version via "ver" IOVAR --- */
 	memset(sc->fw_version, 0, sizeof(sc->fw_version));
-	err = cyw_iovar_get(sc, "ver", sc->fw_version,
+	err = cyw_fil_iovar_data_get(sc, "ver", sc->fw_version,
 	    sizeof(sc->fw_version) - 1);
 	if (err) {
 		device_printf(sc->dev,
